Refactor commands into their own package
Refactoring to accommodate growing scale of the application. An instance
of commands is stood up and populated with the users configuration.
